fix: reverted to /lib folder, added game files
