+++ apps/microgame/src/app/(payload)/layout.tsx
@@ -1,7 +1,7 @@
 /* THIS FILE WAS GENERATED AUTOMATICALLY BY PAYLOAD. */
 /* DO NOT MODIFY IT BECAUSE IT COULD BE REWRITTEN AT ANY TIME. */
 import config from "@payload-config";
-// import "@payloadcms/next/css";
+import "@payloadcms/next/css";
 import type {ServerFunctionClient} from "payload";
 import {handleServerFunctions, RootLayout} from "@payloadcms/next/layouts";
 import React from "react";
