+++ apps/microgame/src/app/(payload)/page.tsx
@@ -1,19 +1,20 @@
 "use client";
 
-import QueryProviderWrapper from "../QueryProviderWrapper";
-import FuelProviderWrapper from "../../components/Terminal/FuelProviderWrapper";
-import Terminal from "../../components/Terminal/Terminal";
+import FuelProviderWrapper from "../../../../../libs/shared/ui/Terminal/FuelProviderWrapper";
+import { QueryClientProvider } from "@tanstack/react-query";
+import { queryClient } from "../../../../../libs/shared/lib/queryClient";
+import Terminal from "../../../../../libs/shared/ui/Terminal/Terminal";
 
-import "./index.css";
+import "../../../../../libs/shared/ui/index.css";
 
 export default function Home() {
   return (
-    <QueryProviderWrapper>
+    <QueryClientProvider client={queryClient}>
       <FuelProviderWrapper>
         <div className="min-h-screen w-full flex items-center justify-center bg-black font-['VT323',monospace]">
           <Terminal />
         </div>
       </FuelProviderWrapper>
-    </QueryProviderWrapper>
+    </QueryClientProvider>
   );
 }
