--- apps/microgame/src/app/QueryProviderWrapper.tsx
@@ -1,11 +1,0 @@
-'use client'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-
-export default function QueryProviderWrapper({ children }: { children: React.ReactNode }) {
-    const queryClient = new QueryClient()
-    return (
-        <QueryClientProvider client={queryClient}>
-            {children}
-        </QueryClientProvider>
-    )
-}
