+++ apps/microgame/tsconfig.json
@@ -27,7 +27,7 @@
     ".next/types/**/*.ts",
     "dist/apps/microgame/types/**/*.ts",
     "next-env.d.ts"
-  ],
+, "../../libs/shared/lib/fuelProvider.ts"  ],
   "exclude": [
     "node_modules"
   ]
