+++ libs/shared/ui/Game/games/blind-fear/audio-off.png
