+++ libs/shared/ui/Game/games/blind-fear/audio-on.png
