+++ libs/shared/ui/Game/games/blind-fear/font.png
