+++ libs/shared/ui/Game/games/blind-fear/game.js
