+++ libs/shared/ui/Game/games/blind-fear/spritesheet.png
