+++ libs/shared/ui/Game/games/blind-fear/spritesheet16.png
