+++ libs/shared/ui/Terminal original/SecretNotes.tsx
@@ -1,10 +1,6 @@
 import { SECRET_NOTES } from '../../lib/constants';
 
-interface SecretNotesProps {
-  onReturn: () => void;
-}
-
-const SecretNotes = ({ onReturn }: SecretNotesProps) => {
+export const SecretNotes = () => {
   return (
     <div className="secret-notes">
       <div className="mb-4 border-b border-terminal-red/50 pb-2">
@@ -14,7 +10,7 @@ const SecretNotes = ({ onReturn }: SecretNotesProps) => {
         </div>
         <p className="text-terminal-blue text-sm font-bold">TOP SECRET // T-REX EXECUTIVE ACCESS // CLEARANCE LEVEL: CEO</p>
       </div>
-
+      
       <div className="space-y-6">
         {SECRET_NOTES.map((note, index) => (
           <div key={index} className={`note p-3 border-l-4 ${note.isHighlighted ? 'border-terminal-red bg-terminal-red/10 text-terminal-red' : 'border-terminal-blue bg-terminal-blue/5'}`}>
@@ -28,7 +24,7 @@ const SecretNotes = ({ onReturn }: SecretNotesProps) => {
           </div>
         ))}
       </div>
-
+      
       <div className="mt-6 border-t border-dashed border-terminal-blue pt-4 flex items-center">
         <div className="w-3 h-3 bg-terminal-green animate-ping mr-2"></div>
         <p className="text-terminal-green font-bold">{"> PRESS [ ENTER ] TO RETURN TO DLM-2000 COMMAND PROMPT"}</p>
@@ -36,5 +32,3 @@ const SecretNotes = ({ onReturn }: SecretNotesProps) => {
     </div>
   );
 };
-
-export default SecretNotes;
