+++ libs/shared/ui/Terminal original/TerminalHeader.tsx
@@ -1,4 +1,4 @@
-const TerminalHeader = () => {
+export const TerminalHeader = () => {
   return (
     <div className="terminal-header h-8 bg-terminal-bg border-b border-terminal-text/30 flex items-center px-4">
       <div className="flex space-x-2 items-center">
@@ -10,5 +10,3 @@ const TerminalHeader = () => {
     </div>
   );
 };
-
-export default TerminalHeader;
