+++ libs/shared/ui/input.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "../../lib/utils"
+import { cn } from "../../shared/utils/cn"
 
 export interface InputProps
   extends React.InputHTMLAttributes<HTMLInputElement> { }
