+++ libs/shared/ui/select.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import * as SelectPrimitive from "@radix-ui/react-select";
 import { Check, ChevronDown, ChevronUp } from "lucide-react";
 
-import { cn } from "../../lib/utils";
+import { cn } from "../../shared/utils/cn";
 
 const Select = SelectPrimitive.Root;
